+++ mailet/standard/src/test/java/org/apache/james/transport/matchers/SenderIsLocalTest.java
@@ -31,6 +31,7 @@ import java.util.Collection;
 import javax.mail.MessagingException;
 
 import org.apache.james.core.MailAddress;
+import org.apache.james.core.MaybeSender;
 import org.apache.mailet.Mail;
 import org.apache.mailet.MailetContext;
 import org.apache.mailet.Matcher;
@@ -106,7 +107,7 @@ public class SenderIsLocalTest {
         //Given
         Mail mail = FakeMail.builder()
             .name("mail")
-            .sender(MailAddress.nullSender())
+            .sender(MaybeSender.nullSender())
             .recipient(ANY_AT_JAMES)
             .build();
         //When