+++ mailet/standard/src/test/java/org/apache/james/transport/matchers/SenderIsTest.java
@@ -28,6 +28,7 @@ import java.util.Optional;
 import javax.mail.MessagingException;
 
 import org.apache.james.core.MailAddress;
+import org.apache.james.core.MaybeSender;
 import org.apache.mailet.base.test.FakeMail;
 import org.apache.mailet.base.test.FakeMatcherConfig;
 import org.junit.jupiter.api.BeforeEach;
@@ -72,7 +73,7 @@ class SenderIsTest {
         FakeMail fakeMail = FakeMail.builder()
             .name("mail")
             .recipient(recipient)
-            .sender(MailAddress.nullSender())
+            .sender(MaybeSender.nullSender())
             .build();
 
         assertThat(matcher.match(fakeMail)).isNull();
@@ -103,7 +104,7 @@ class SenderIsTest {
         FakeMail fakeMail = FakeMail.builder()
             .name("mail")
             .recipient(recipient)
-            .sender(MailAddress.nullSender())
+            .sender(MaybeSender.nullSender())
             .build();
 
         assertThat(matcher.match(fakeMail)).containsExactly(recipient);