+++ protocols/smtp/src/main/java/org/apache/james/protocols/smtp/MailEnvelope.java
@@ -50,33 +50,15 @@ public interface MailEnvelope {
     List<MailAddress> getRecipients();
 
     /**
-     * Return the sender of the mail which was supplied int the MAIL FROM:
-     * command. If its a "null" sender, null will get returned
-     *
-     * @deprecated @see {@link #getMaybeSender()}
-     *
-     * Note that SMTP null sender ( "&lt;&gt;" ) needs to be implicitly handled by the caller under the form of 'null' or
-     * {@link MailAddress#nullSender()}. Replacement method adds type safety on this operation.
-     *
-     * @return sender
-     */
-    @Deprecated
-    default MailAddress getSender() {
-        return getMaybeSender().asOptional().orElse(MailAddress.nullSender());
-    }
-
-    /**
      * Returns the sender of the message, as specified by the SMTP "MAIL FROM" command,
      * or internally defined.
      *
-     * 'null' or {@link MailAddress#nullSender()} are handled with {@link MaybeSender}.
+     * 'null' is handled with {@link MaybeSender}.
      *
      * @since Mailet API v3.2.0
      * @return the sender of this message wrapped in an optional
      */
-    default MaybeSender getMaybeSender() {
-        return MaybeSender.of(getSender());
-    }
+    MaybeSender getMaybeSender();
 
     /**
      * Return the OutputStream of the message